+++ src/openhems/modules/energy_strategy/emhass_strategy.py
@@ -144,5 +144,5 @@ class EmhassStrategy(EnergyStrategy):
 				switchOnRate = 100 * ( np.dot(vals, rates) ) / value
 			deferable = self.deferables[deferableName]
 			doSwitchOn = self.evaluatePertinenceSwitchOn(switchOnRate, deferable.node)
-			self.switchOnSchedulable(deferable.node, cycleDuration, doSwitchOn)
+			self.switchSchedulable(deferable.node, cycleDuration, doSwitchOn)
 		return True