+++ src/openhems/modules/util/cast_utility.py
@@ -35,7 +35,7 @@ class CastUtililty:
 				raise CastException("Unavailable value for '"+value+"'", 0)
 			retValue = int(float(value)) # int("0.0") crash while float("0") don't...
 		else:
-			raise CastException("Impossible cast to int: Undefined algorythm : '"+type(value)+"'", 0)
+			raise CastException("Impossible cast to int: Undefined algorythm : '"+str(type(value))+"'", 0)
 		return retValue
 	@staticmethod
 	def toTypeBool(value):
@@ -129,7 +129,7 @@ class CastUtililty:
 			else:
 				raise CastException("Incorect string value for  float: '"+value+"'", 0)
 		else:
-			raise CastException("Impossible cast to list: Undefined algorythm : '"+type(value)+"'", 0)
+			raise CastException("Impossible cast to list: Undefined algorythm : '"+str(type(value))+"'", 0)
 		return retValue
 
 	@staticmethod