+++ src/openhems/server.py
@@ -56,8 +56,10 @@ class OpenHEMSServer:
 				self.logger.critical(msg)
 				throwErr = msg
 		if throwErr is not None:
+			self.logger.error(str(throwErr))
 			raise ConfigurationException(throwErr)
 		self.allowSleep = len(self.strategies)==1
+		self.inOverLoadMode = False # in over load mode, we have node deactivate for safety
 
 	def getSchedule(self):
 		"""
@@ -73,6 +75,46 @@ class OpenHEMSServer:
 					schedule[myid] = sc
 		return schedule
 
+	def check(self):
+		""""
+		For safety, Avoid over-load, check margin power.
+		If margin power is used, we switch off devices
+		"""
+		marginPowerOn = self.network.getMarginPowerOn()
+		self.logger.debug("Security margin power:%s",marginPowerOn)
+		if marginPowerOn<0: # Need to switch off (deactivate) some nodes
+			self.logger.warning(
+				"Margin power On is negativ (%f): Need to sitch off devices.",
+				marginPowerOn)
+			elems = self.network.getAll("out")
+			elems.sort( # start from the less priority
+				key=lambda x:x.getPriority()
+			)
+			for elem in elems:
+				if marginPowerOn<0 and elem.isSwitchable() and elem.isOn():
+					power = elem.getCurrentPower()
+					self.logger.info("Switch off '%s' due to missing power margin.", elem.id)
+					if elem.switchOn(False):
+						self.logger.error("Fail switch off '%s' due to missing power margin.", elem.id)
+					else:
+						marginPowerOn += power
+						elem.setActivate(False)
+						self.inOverLoadMode = True
+		elif self.inOverLoadMode and marginPowerOn>0: # Try to re-activate nodes
+			inOverLoadMode = False
+			# start from the bigest priority (default order)
+			elems = self.network.getAll("out")
+			for elem in elems:
+				if not elem.isActivate():
+					if elem.getMaxPower()>marginPowerOn: # re-activate the node
+						elem.setActivate(True)
+						# Do just one at each loop for safety
+						return marginPowerOn
+					inOverLoadMode = True
+			self.inOverLoadMode = inOverLoadMode
+		return marginPowerOn
+
+
 	def loop(self, loopDelay, now=None):
 		"""
 		It's the content of each loop.
@@ -82,6 +124,7 @@ class OpenHEMSServer:
 			now = datetime.datetime.now()
 		self.logger.debug("OpenHEMSServer.loop()")
 		self.network.updateStates()
+		self.check()
 		time2wait = 86400
 		allowSleep = self.allowSleep and loopDelay>LOOP_DELAY_VIRTUAL
 		for strategy in self.strategies: